fix: support for elife reference articles.
With the library in its default state it was not possible to open ay of the
elife reference articles.

The first issue was caused by the post-transformation validator, where once
the XML has been transformed from JATS to Textures JATS format there were
some errors. To workaround this, I've added a new flag that when set will
ignore validation failures and continue to load the article.

The second failure was caused by citations with unsupported 'publication_type'
attributes. I've updated the matcher with new rules for some types, e.g. to
map elife's 'web' type to 'webpage'. And for types where no current match was
obvious I've added some logic to default to the type 'journal'.

These fixes are enough to get Texture to now load and display articles. But
further work will be needed to fix these problems properly. I've raised issues