fix: display references as a bulleted list
Updated the Reference component to use a bullet delimiter rather than a
number when displayed on the main view. The solution is far from ideal,
where it would've been preferable to update the getLabel() method to take
the mode as input and return the correct label based on state. But there
are other points in the code that ignore the getLabel() method also so the
extent of the refactoring would be a little too much for the scope of this
ticket.

In addition, I also made the following tweaks to the Reference component as
per the design requirements.
- Removed the label when a reference is displayed in preview mode.
- Added placeholder 'edit' and 'delete' buttons, but these will be wired up
  at a later date.
- Added some documentation to various functions.