fix: references should be ordered alphabetically
Textures default behaviour was to sort based on citation order. I've
since broken the sorting code out into some helper functions, and
implemented a new method that sorts by name, falling back to year if
the names are identical.
The fix is suitable for now, but further work should be done to make
the sorting configurable.