fix: chapter references should use 'pp' rather than 'p'
In chapter references, if both the fpage and lpage elements have a
value then the editor should display 'pp' rather than 'p'. If only
the fpage is set, then it should use 'p'.
In addition, there also should not be a space between 'pp'/'p' and
the value of fpage.